LayoutTests/http/tests/security/contentSecurityPolicy/object-src-param-src-blocked-expected.txt

CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/plugins/resources/mock-plugin.pl?src because it does not appear in the object-src directive of the Content Security Policy.
This test passes if there is a console message saying the plugin was blocked.
